+++ CHANGELOG.md
@@ -30,6 +30,8 @@ in 8.x, 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 
 ## Lab
 
+### 2026-09-24
+ * Improved Smart Build Module Tool **(DEV team only)**
 ### 2026-09-23
  * Added Smart Build Module Tool **(DEV team only)**
  * Updated PackingSolver dependency